refactor for hw2
