+++ .github/workflows/pylint.yml
@@ -1,6 +1,10 @@
 name: pylint_my_project
 
-on: [push, pull_request]
+on:
+  push:
+    branches: [ homework1 ]
+  pull_request:
+    branches: [ homework1 ]
 
 jobs:
   build:
@@ -17,10 +21,9 @@ jobs:
       - name: Install dependencies
         working-directory: ./ml_project
         run: |
-          python -m pip install --upgrade pip
-          python -m pip install -r requirements.txt
+          pip install .
 
       - name: Pylint
         working-directory: ./ml_project
         run: |
-          pylint src train.py predict.py  --ignore=tests --disable=C0114,C0115,C0116 --fail-under=7.0
+          pylint ml_project --disable=C0114,C0115,C0116 --fail-under=7.0
