+++ .github/workflows/test_ml_project.yml
@@ -21,10 +21,9 @@ jobs:
       - name: Install dependencies
         working-directory: ./ml_project
         run: |
-          python -m pip install --upgrade pip
-          python -m pip install -r requirements.txt
+          pip install .
 
       - name: Testing
         working-directory: ./ml_project
         run: |
-          python -m pytest . -v
+          pytest -v .
