+++ ml_project/README.md
@@ -7,7 +7,7 @@ Download dataset from [Kaggle](https://www.kaggle.com/ronitf/heart-disease-uci) 
 * `pip3 install virtualenv`
 * `virtualenv ml_project_venv` 
 * `source ml_project_venv/bin/activate`
-* `pip install -r requirements.txt`
+* `pip install .`
 
 ## EDA in jupyter notebook
 ```bash
@@ -17,18 +17,23 @@ jupyter notebook notebooks/heart_eda.ipynb
 ## Training
 
 ```bash
-python -m train --config-name ./configs/model_1.yml
-python -m train --config-name ./configs/model_2.yml
+python ml_project/train.py --config-name ./configs/model_1.yml
+python ml_project/train.py --config-name ./configs/model_2.yml
 ```
 
 ## Prediction
 
 ```bash
-python -m predict --config-name ./configs/model_1.yml
-python -m predict --config-name ./configs/model_2.yml
+python ml_project/predict.py --config-name ./configs/model_1.yml
+python ml_project/predict.py --config-name ./configs/model_2.yml
 ```
 
 ## Local project testing
 ```bash
-python -m pytest -v .
+pytest -v .
+```
+
+## Check code style with pylint
+```bash
+pylint ml_project --disable=C0114,C0115,C0116 --fail-under=7.0
 ```
